updates unit testing to match stats changes
