+++ tests/unit/__mocks__/stats.js
@@ -1,0 +1,9 @@
+export const stats = {
+  "Jan 6": { started: 0, completed: 0, interruptions: 0 },
+  "Jan 7": { started: 0, completed: 0, interruptions: 0 },
+  "Jan 8": { started: 100, completed: 100, interruptions: 0 },
+  "Jan 9": { started: 200, completed: 200, interruptions: 0 },
+  "Jan 10": { started: 300, completed: 210, interruptions: 3 },
+  "Jan 11": { started: 200, completed: 120, interruptions: 2 },
+  "Jan 12": { started: 100, completed: 80, interruptions: 0 },
+};
